Extract request for creating a receipt into a service

+++ src/services/deleteReport.js
@@ -1,8 +1,9 @@
 import axios from 'axios'
 import {config, baseURL} from './axiosConfig'
 
-export default async report_id => {
+export default async (report_id, mockError) => {
   try{
+    if(mockError) throw 'mockingError'
     const company_id = localStorage.getItem('company_id')
     const response = await axios.delete(`${baseURL}/${company_id}/report/${report_id}`, config)
     return response.data
